Demo error of PositionSource in QML on Sailfish
When the PostionSource start() method is called in QML,  the following
error is generated
QMetaMethod::invoke: Unable to handle unregistered datatype
'QGeoPositionInfo'

This throwaway project demos this both with a standard import of
QtPositioning 5.2
and with equivalent code "in-line" called by importing PositionSource
5.2
Switching between these 2 imports by commenting in / out allow
switching from the standard import and the "in-line" equivalent.

The idea of the "in-line" code is to allow tinkering and thus possibly
to track down the source of the error (and suggest a solution).

Copyright of the inline src and headers files "pinched" from
QtPositioning remains that of Diga, as stated at the top of the files.
They are included here for demonstration purposes only.